Update README.md
Hangman Game
Hangman is a word-guessing game where one player (computer) thinks of a word, and the other player tries to guess it by suggesting letters. Each incorrect guess results in a part of a hangman being drawn( life deducted) the player is provided with finite number of lives, and the game ends when either the word is guessed correctly or when the player runs out of lives.